+++ libpm/command_tests.py
@@ -120,7 +120,7 @@ class CommandTests:
 
             _urls = {
                 "http://localhost:8080/api/v1/my-ip": 200,
-                "http://localhost:8080/flag?teamid=t02&flag=" + UtilsTests.random_flag(): 403,
+                "http://localhost:8080/flag?team_id=t02&flag=" + UtilsTests.random_flag(): 403,
                 "http://localhost:8080/api/v1/teams": 200,
                 "http://localhost:8080/api/v1/services": 200,
                 "http://localhost:8080/api/v1/scoreboard": 200,